pyproject.toml
[tool.poetry]
name = "mnist"
version = "0.1.0"
description = "Predict numbers from 0 to 9"
authors = ["Yaroslav Kliukin <kliukin.iad@phystech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
torch = {version = "^2.0.1+cpu", source = "pytorch"}
torchvision = {version = "^0.15.2+cpu", source = "pytorch"}
pandas = "^2.1.1"
dvc = {extras = ["gdrive"], version = "^3.27.0"}
hydra-core = "^1.3.2"
lightning = "^2.1.2"
torchmetrics = "^1.2.1"
mlflow = "^2.8.1"
onnxruntime = "^1.16.3"
onnx = "^1.15.0"
fire = "^0.5.0"

[tool.poetry.dev-dependencies]
pre-commit = "^3.4.0"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
src_paths = ["mnist"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
